requirements.txt
googlesearch-python
pymupdf
google-generativeai
python-dotenv
langchain-google-genai
langchain_core
streamlit
requests